Corrected the warp provider implementation for Java 1.6.
