+++ arquillian-warp-spring/src/test/java/org/jboss/arquillian/warp/extension/spring/container/provider/ProviderBaseTestCase.java
@@ -26,6 +26,7 @@ import org.springframework.web.servlet.HandlerInterceptor;
 import org.springframework.web.servlet.ModelAndView;
 
 import javax.servlet.ServletRequest;
+import javax.servlet.http.HttpServletRequest;
 
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
@@ -82,10 +83,10 @@ public class ProviderBaseTestCase {
     @SuppressWarnings("unchecked")
     protected void setSpringMvcResult() throws Exception {
 
-        ServletRequest mockServletRequest = mock(ServletRequest.class);
+        HttpServletRequest mockServletRequest = mock(HttpServletRequest.class);
         when(mockServletRequest.getAttribute(Commons.SPRING_MVC_RESULT_ATTRIBUTE_NAME)).thenReturn(springMvcResult);
 
-        Instance<ServletRequest> mockServletRequestInstance = mock(Instance.class);
+        Instance<HttpServletRequest> mockServletRequestInstance = mock(Instance.class);
         TestReflectionHelper.setFieldValue(instance, "servletRequestInstance", mockServletRequestInstance);
         when(mockServletRequestInstance.get()).thenReturn(mockServletRequest);
     }
